Consolidated weapon bay printing with standard weapon and equipment layout.
